Play loaded events
There seems to be a bug with is_mouse_button_released when working with
replayed events. is_mouse_button_released won't trigger for a replay
that would cause is_mouse_button_released to be triggered outside of a
replay. It looks like the previous state when a mouse button is released
isn't updated in Raylib (See
https://github.com/raysan5/raylib/blob/9c2ba3bfb7166b71a852d4a4912363d0977c8676/src/rcore.c#L2701)